Migrate restaurant feed to a separate component
